requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies for advanced features
asyncio>=3.4.3
aiohttp>=3.8.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0